Move get_s3_connection to utils module
